+++ src/cli/flatcc_cli.c
@@ -23,6 +23,7 @@ void usage(FILE *fp)
             "  -r, --recursive            Recursively generate included schema files\n"
             "  -a                         Generate all (like -cwvr)\n"
             "  -g                         Use _get suffix only to avoid conflicts\n"
+            "  -s                         Silence warnings\n"
             "  -d                         Dependency file like gcc -MMD\n"
             "  -I<inpath>                 Search path for include files (multiple allowed)\n"
             "  -o<outpath>                Write files relative to this path (dir must exist)\n"
@@ -82,6 +83,11 @@ void help(FILE *fp)
         "only 'Monster_name_get(monster)` will be generated and not also\n"
         "'Monster_name(monster)'. This avoids potential conflicts with\n"
         "other generated symbols when a schema change is impractical.\n"
+        "\n"
+        "-s Silence warnings, notably on potential conflicts that might cause\n"
+        "C source code to generate errors. These warnings are not perfect and can\n"
+        "be ignored if the C code compiles, though it is recommended to change\n"
+        "names in the schema if possible.\n"
         "\n"
         "-d generates a dependency file, e.g. 'monster.fbs.d' in the output dir.\n"
         "\n"
@@ -343,6 +349,9 @@ int set_opt(flatcc_options_t *opts, const char *s, const char *a)
         return noarg;
     case 'g':
         opts->cgen_no_conflicts = 1;
+        return noarg;
+    case 's':
+        opts->silence = 1;
         return noarg;
     case 'd':
         opts->gen_dep = 1;